Use caching if seed is fixed.

+++ nengo/cache.py
@@ -18,8 +18,12 @@ logger = logging.getLogger(__name__)
 class DecoderCache(object):
     _DECODER_EXT = '.npy'
     _SOLVER_INFO_EXT = '.pkl'
+    DEFAULT_DIR = os.path.expanduser(os.path.join('~', '.nengo-cache'))
 
-    def __init__(self, cache_dir):
+    def __init__(self, read_only=False, cache_dir=None):
+        self.read_only = read_only
+        if cache_dir is None:
+            cache_dir = self.DEFAULT_DIR
         self.cache_dir = cache_dir
         if not os.path.exists(self.cache_dir):
             os.mkdir(self.cache_dir)
@@ -64,9 +68,10 @@ class DecoderCache(object):
                 logger.info("Cache miss [{0}].".format(key))
                 decoders, solver_info = solver(
                     activities, targets, rng=rng, E=E)
-                np.save(decoder_path, decoders)
-                with open(solver_info_path, 'wb') as f:
-                    pickle.dump(solver_info, f)
+                if not self.read_only:
+                    np.save(decoder_path, decoders)
+                    with open(solver_info_path, 'wb') as f:
+                        pickle.dump(solver_info, f)
             return decoders, solver_info
         return cached_solver
 
@@ -97,3 +102,17 @@ class DecoderCache(object):
 
     def _get_solver_info_path(self, key):
         return os.path.join(self.cache_dir, key + self._SOLVER_INFO_EXT)
+
+
+class NoDecoderCache(object):
+    def wrap_solver(self, solver):
+        return solver
+
+    def get_size(self):
+        return 0
+
+    def shrink(self, limit=0):
+        pass
+
+    def invalidate(self):
+        pass
